+++ packages/us-web-util/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.2.0
+
+- Fix TypeScript errors for `invariant`.
+- Add `scrollToBottom`, `scrollToTop`
+
 # 1.1.0
 
 - Export `js-cookie` library as 'blessed' solution.